step1 - 수정사항 수정
